requirements.txt
setuptools
pandas
lxml
selenium
undetected-chromedriver
google-cloud-storage
google-auth